+++ src/Fields/Group.php
@@ -1,20 +1,13 @@
 <?php
 
-namespace Towa\Acf\Fields;
+namespace DigitOne\Acf\Fields;
 
-use Towa\Acf\BaseField;
+use DigitOne\Acf\BaseParentField;
 
-class Group extends BaseField
+class Group extends BaseParentField
 {
     protected $prefix;
     protected $name = 'group';
     protected $label = 'Group';
     protected $type = 'group';
-
-    public function build(array $parameter = [])
-    {
-        $default = [ 'sub_fields' => [] ];
-
-        return parent::build(array_merge($default, $parameter));
-    }
 }